Expose search subquery conditions
We'd like to use pg_search inside an Arel pipeline, and need finer
control over joining and sorting. This allows us to access the filter
conditions alone.

A better fix here would be to expose all of the relevant query fragments
as Arel nodes as part of the API, but this is Good Enough for Me™.